Add a series of negative tests for incorrect input
This required a relatively large rearrangement to the testing
infrastructure to break out detecing problems from failing the testing
package directly.

In addition, the default fail mode has been switched to Fail() from
Fatal(), as so multiple astRegressions can be run in one test.  This
is easier to read, even if it sacrifices a potential parallelization
opportunity sometimes.

Signed-off-by: Daniel Farina <daniel@fdr.io>