Do not indent on account of pointer rendering
This is to make test output more stable under change (it does not
affect the determinism of testing).  It causes a large change in
output of the current regression tests, but there is no functional
change in any non-test code.

Previously, a change in the level of indirection of a field in a
struct is changed (inlined, single-pointer, double-pointer, etc) would
cause a lot of whitespace alterations.  No longer.

Signed-off-by: Daniel Farina <daniel@fdr.io>