Change attributes supported in dogconf
On reflection, there is no need to know anything about usernames and
passwords.  Even though they are provided by the startup packet, it
seems as though database name routing is all that is needed for now.
And, to take care of replicas, one must support rewriting distinct
database names to one database name, so add this feature.

Signed-off-by: Daniel Farina <daniel@fdr.io>